pyproject.toml
[tool.poetry]
name = "kvartirka_test"
version = "0.1.0"
description = ""
authors = ["Irek Gimatov <gimatov@list.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
djangorestframework-simplejwt = "^4.7.2"
django-treebeard = "^4.5.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
django-debug-toolbar = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
